Enhanced article rendering with markdown support
In this session, we focused on refactoring the artifact detail page to improve its content rendering capabilities. The main changes involved extracting the content rendering logic into dedicated components, with ArticleRenderer being moved to its own module and enhanced with proper Markdown support using react-markdown and related plugins. This separation of concerns allows for better maintainability and potential reuse of content rendering across the application.
We also made several UX improvements to the artifact detail page layout, including relocating the network badge beneath the title, restructuring the media sources section to appear above content toggles, and removing artificial height constraints on rendered content. These changes make the content more accessible and readable while maintaining a clean, organized interface that better matches the intended design of a modern content viewing experience.
