+++ src/components/ContentRenderers/index.js
@@ -1,0 +1,5 @@
+// src/components/ContentRenderers/index.js
+export { default as ArticleRenderer } from './ArticleRenderer';
+// We'll uncomment these once they're implemented
+// export { default as AudioRenderer } from './AudioRenderer';
+// export { default as VideoRenderer } from './VideoRenderer';
